requirements.txt
Flask==2.0.1
python-dotenv==0.19.0
sendgrid==6.9.7
Flask-WTF==0.15.1
plotly==5.18.0
requests==2.31.0
pandas==2.2.1
openpyxl==3.1.2 